feat: add user authentication handling in category actions and layout for improved security

+++ app/(protected)/budgets/categories/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { createClient } from '@/lib/supabase/server';
+import { createClient, getUser } from '@/lib/supabase/server';
 import { ICategory, ICategoryPageParams } from "./types";
 import { revalidatePath } from 'next/cache';
 
@@ -12,9 +12,12 @@ export async function getCategories(params: ICategoryPageParams): Promise<ICateg
     sortOrder = "asc" 
   } = params;
 
+  const user = await getUser();
+
   let query = supabase
     .from("categories")
-    .select("id, name");
+    .select("id, name")
+    .eq("user_id", user?.id);
 
   if (search) {
     query = query.ilike("name", `%${search}%`);
@@ -34,10 +37,11 @@ export async function getCategories(params: ICategoryPageParams): Promise<ICateg
 
 export async function createCategory(name: string){
   const supabase = await createClient();
+  const user = await getUser();
 
     const { error } = await supabase
     .from('categories')
-    .insert([{ name }]);
+    .insert([{ name, user_id: user?.id }]);
 
   // TODO handle error properly
   if (error) {
@@ -50,11 +54,13 @@ export async function createCategory(name: string){
 
 export async function updateCategory(id: string, name: string){
   const supabase = await createClient();
+  const user = await getUser();
 
   const { error } = await supabase
     .from('categories')
     .update({ name })
-    .eq('id', id);
+    .eq('id', id)
+    .eq('user_id', user?.id);
 
   // TODO handle error properly
   if (error) {
@@ -67,11 +73,13 @@ export async function updateCategory(id: string, name: string){
 
 export async function deleteCategory(id: string){
   const supabase = await createClient();
+  const user = await getUser();
 
   const { error } = await supabase
     .from('categories')
     .delete()
-    .eq('id', id);
+    .eq('id', id)
+    .eq('user_id', user?.id);
 
   // TODO handle error properly
   if (error) {
